Big cleanup. Every services is now seperated with cleaner dependencies
