+++ jeto/__init__.py
@@ -22,25 +22,25 @@ babel = Babel(app)
 oauth = OAuth()
 principal = Principal(app)
 
-google = oauth.remote_app('google',
-                          base_url='https://www.google.com/accounts/',
-                          authorize_url='https://accounts.google.com/o/oauth2/auth',
-                          request_token_url=None,
-                          request_token_params={'scope': 'openid profile email',
-                                                'response_type': 'code'},
-                          access_token_url='https://accounts.google.com/o/oauth2/token',
-                          access_token_method='POST',
-                          access_token_params={'grant_type': 'authorization_code'},
-                          consumer_key=app.config['GOOGLE_CLIENT_ID'],
-                          consumer_secret=app.config['GOOGLE_CLIENT_SECRET'])
+google = oauth.remote_app(
+    'google',
+    base_url='https://www.google.com/accounts/',
+    authorize_url='https://accounts.google.com/o/oauth2/auth',
+    request_token_url=None,
+    request_token_params={
+        'scope': 'openid profile email',
+        'response_type': 'code'
+    },
+    access_token_url='https://accounts.google.com/o/oauth2/token',
+    access_token_method='POST',
+    access_token_params={'grant_type': 'authorization_code'},
+    consumer_key=app.config['GOOGLE_CLIENT_ID'],
+    consumer_secret=app.config['GOOGLE_CLIENT_SECRET']
+)
 
 lm = LoginManager(app)
 
 import jeto.core
-import jeto.models.permission
-import jeto.models.vagrant
-import jeto.models.team
-import jeto.models.host
-import jeto.models.project
-import jeto.models.domainController
 import jeto.controllers
+from jeto.models import *
+from jeto.services import *
