+++ jeto/services/htpasswords.py
@@ -1,0 +1,95 @@
+import requests as req
+from flask import request, json
+
+from flask.ext.restful import Resource, fields, marshal
+from jeto import app
+
+
+htpassword_list_fields = {
+    'slug': fields.String,
+    'users': fields.List(fields.String),
+}
+
+
+class HtpasswordService(object):
+    def _get_url(self):
+        return 'http://' + app.config.get('HTPASSWORD_API_URL') + ':' +\
+            app.config.get('HTPASSWORD_API_PORT')
+
+    def _get_headers(self):
+        return {
+            "Content-Type": "application/json",
+        }
+
+
+class HtpasswordApi(Resource, HtpasswordService):
+    def get(self):
+        r = req.get(self._get_url())
+        items = r.json()['lists']
+
+        return {
+            'lists': map(lambda t: marshal(t, htpassword_list_fields), items),
+        }
+
+    def post(self, slug=None):
+        name = request.json['name']
+
+        data = json.dumps({'name': name})
+        # Should mean we are adding a new user
+        r = req.post(self._get_url(),
+                     headers=self._get_headers(),
+                     data=data)
+        content = r.content
+
+        return content
+
+    def delete(self, slug):
+        url = self._get_url() + '/{}'.format(slug)
+        r = req.delete(url=url, headers=self._get_headers())
+        return r.content
+
+    def put(self, slug=None):
+        domain = request.json['domain']
+        ip = request.json['ip'].strip()
+        data = json.dumps({'site': domain, 'ip': ip})
+        r = req.put(self._get_url() + '/{}'.format(slug),
+                    headers=self._get_headers(),
+                    data=data)
+
+        return r.content
+
+
+class HtpasswordListApi(Resource, HtpasswordService):
+    def get(self, slug):
+        r = req.get(self._get_url(slug))
+        htpassword = r.json()
+        return {'item': htpassword}
+
+    def delete(self, slug):
+        r = req.delete(self._get_url(slug))
+        return r.content
+
+    def put(self, slug):
+        users = request.json['users']
+        for user in users:
+            if 'state' in user:
+                if user['state'] == 'DELETE':
+                    req.delete(self._get_url(slug) +
+                               '/{}'.format(user['username']))
+
+                if user['state'] == 'CREATE':
+                    data = json.dumps({
+                        'username': user['username'],
+                        'password': user['password']
+                    })
+                    req.post(self._get_url(slug),
+                             headers=self._get_headers(), data=data)
+
+        return self.get(slug)
+
+    def _get_url(self, slug):
+        return super(HtpasswordListApi, self)._get_url() + '/{}'.format(slug)
+
+    def _get_headers(self):
+        return {'Accept': 'application/json',
+                'Content-Type': 'application/json'}
