+++ requirements.txt
@@ -28,6 +28,6 @@ six==1.6.1
 speaklater==1.3
 wsgiref==0.1.2
 redis==2.9.1
-rq
+rq=0.5.3
 times==0.6.2
 ansiconv==1.0.0
